requirements.txt
streamlit
python-dotenv
gtts
pygame
pydub
speechrecognition
groq
ffmpeg-python
